accidents replaced with reports
